refactor: extract 'withdraw to btc onchain address' response data
